Move FF check into WooPosGetProductById use case
