+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/woopos/home/variations/WooPosVariationsViewModelTest.kt
@@ -772,8 +772,6 @@ class WooPosVariationsViewModelTest {
             mapper,
             localCatalogSyncRepository,
             selectedSite,
-            productMapper,
-            localCatalogStore,
         ).let {
             it.init(productId, sourceType = sourceType)
             it
